+++ sdk/communication/AzureCommunicationCommon/Source/Authentication/StaticTokenCredential.swift
@@ -24,9 +24,6 @@
 //
 // --------------------------------------------------------------------------
 
-#if canImport(AzureCore)
-    import AzureCore
-#endif
 import Foundation
 /**
  The Azure Communication Services User token credential.